Normalise annotations on AnnotatedIOBuffer read
It is easy to come across situations where content with the same styling
is consecutively written to an AnnotatedIOBuffer, resulting in what
could be one region with a certain style being represented with many
identical regions instead. This is entirely valid, and handled
correctly, but it is much nicer in a number of respects if any such
regions are consolidated.

The annotatedstring_optimize! function does exactly this, and so we
apply it to the AnnotatedString result when reading from an
AnnotatedIOBuffer.