opaque_closure: Lookup optimized oc source inside code instance
This is an alternative to #53852. I don't think it's semantically legal
to put optimized source into the :source field of a method, but it
should be fine to just look it up from the CodeInstance. That said,
this is a bit of an unusual configuration. In particular it wasn't
even reachable with the surface APIs, which assumed that inferred
IR was always supposed to be compiled.