Locate top module for object in image
Julia objects can be perma-allocated in package images,
each package image has a corresponding top module.

We currently can map Julia objects to loaded image,
but we don't keep track of the corresponding top module.

This can be useful to ask for the build_id of the package
image we are using.

Non perma-allocated objects are mapped to `Main`.