Add `zero` for `Base.TwicePrecision` (#53787)
Since `zero` exists for a `TwicePrecision` type, it makes sense for a
method to exist for an instance as well.
Fixes #52713, which was a regression from v1.9.1. After this, the
following may be displayed without errors:
```julia
julia> r = 1.0*(1:5) .+ im
1.0 + 1.0im:Base.TwicePrecision{Float64}(1.0, 0.0):5.0 + 1.0im
```
However, in this case, the step is a `Base.TwicePrecision`, which seems
to be an implementation detail that's leaking out, although addressing
this may be a separate PR.