TOML: Improve type-stability of BigInt/UInt support
From a type-stability perspective, this restores a lot of our behavior
before #47903. As it turns out, 10 of the 11 uses of `parse_int` introduced
in that PR are unnecessary since the TOML format already requires the
parsed value to be within a very limited range.

Note that this change does not actually revert any functionality (in
contrast to #49576)