Explicit namespaces in sort-related docstrings (#53968)
This represents the unexported names by the fully qualified namespace,
which makes it simpler to interpret the keyword arguments. Otherwise,
for example, it's not obvious where `Forward` comes from in the `sort`
docstring.

Also, replaces some occurrences of `instance <: Algorithm` by `instance
isa Algorithm`, which is the correct relation. This is also closer to
English, which helps with readability.

---------

Co-authored-by: Lilith Orion Hafner <lilithhafner@gmail.com>