Clean up pass system a bit
Return contexts from pass functions, in addition to the expression so
callers of a pass don't need to figure out how to create the context for
that pass. Feels a bit better, but still unclear that this is great.

Also be a bit more careful about reparenting SyntaxTree to the new graph
as attributes are added by passes. This still feels somewhat dubious
as any old SyntaxList data structures attached as attributes will not
see the added attributes.