irinterp: Move irinterp va processing
Currently we perform the va transform inside the innermost construct
of IRInterpretationState. However, I think it makes more sense to
lift this processing one level, so that the `argtypes` for the
constructor that takes `ir` matches `ir.argtypes`, while the one
for the constructor that takes `mi` is the full expanded out
argtypes. NFC for base, but the `ir` constructor is used in
downstream external absints. However, I think this way around
is better for them also, since it's not always clear what format
the argtypes they have are in and the va direction is easy and
well supported, so it's better to have the option.