set `@max_methods 2` for `error`
It is important to always be able to infer the return type of `error`
as `Union{}`, but there's a hitch when a package globally sets
`@max_methods 1` and it causes inference for `error(::Any)` to fail
(JuliaLang/julia#54029).

To address this, this commit sets `@max_methods 2 function error end`
at the definition of `error`. This overrides any global
`@max_methods 1` settings on package side, guaranteeing that return type
inference on `error` is successful always.