inlining: Use tfunc for inserted `isa`/`and` calls
The motivation for this is the same as #54105, where the return types
being filled in by inference can in general be imprecise when running
the compiler with an extended lattice.

This PR is very incomplete - There are at least 6+ more places in
`inlining.jl` where we hard-code return types that will need similar
treatment.