post-opt: taint :consistent when statement may raise inconsistently
When an inconsistent statement doesn’t affect the return value, post-opt
analysis will try to refine it to `:consistent`. However this refinement
is invalid if the statement could throw, as `:consistent` requires
consistent termination.

For the time being, this commit implements the most conservative fix.
There might be a need to analyze `:nothrow` in a data-flow sensitive
way in the post-opt analysis as like we do for `:consistent`.

- closes #53613