inlining: Use `tmeet` on inserted `PiNode` types for extended lattices (#54105)
For extended lattice elements, the type of the resulting PiNode can end
up imprecise when `tmeet(argextype(argex), mft) ⊑ mft`

That said, it's not entirely clear to me what our `IR_FLAG_REFINED`
obligations are in the first place. Are we supposed to mark any
statements that are widened past what inference would have seen? If so,
it seems like we fall far short of that.