cfg_simplify: Don't accidentally consider `GotoIfNot` a throwing terminator (#54262)
This addresses a bug in #54216, where a `GotoIfNot` was accidentally
considered a throwing terminator. Just as I was about to PR this, I
noticed that #54260 had already been opened for the same issue. However,
there's three differences in this PR that made me open it anyway:

1. There's an additional missing case where the terminator is `nothing`
which has special case semantics of allowing any type on it, because it
serves as a deletion marker.

2. My test also test the `EnterNode` and `:leave` terminators, just to
have a complete sampling.

3. I like the code flow in this version slightly better.

Co-authored-by: Cody Tapscott <cody.tapscott@juliahub.com>