irverify: Force terminators to have `Any` type (#54279)
We were mostly consistent about assigning `Any` type to terminators in
absint, but some optimization passes preferred `Union{}` instead. Yet
elsewhere, we accidentally put `Const(EnterNode(...))` here. None of
this really matters, because we're never supposed to look at this, but
it's nice to be able to rely on `Union{}` reliably meaning that control
terminates after the statement (other than for PhiNode, but that's a
separate story) to avoid issues like #54262.