IncrementalCompact: Don't corrupt CFG when finishing early
Finishing an IncrementalComapact before iterating the whole underlying
IRCode is not something that we really do in base, or at least not in
places other than a basic block boundary. However, it can be useful
for downstream consumers, so make it work anyway. Also create a
separate test file for IncrementalCompact tests (of which we really
should have more) and move one existing test.