Support riscv64 as binary platform (#54011)
This patch adds support for `riscv64` as binary platform.

Note: This only allows Julia to handle `riscv64` in `BinaryPlatforms`
and `CPUID`. There is no support for either code generation or building
Julia on RISC-V systems. This is only a necessary first step towards
these goals. In particular, this is necessary to support building
external libraries for RISC-V via BinaryBuilder as described in
https://github.com/JuliaPackaging/BinaryBuilderBase.jl/pull/374.

---------

Co-authored-by: Oscar Smith <oscardssmith@gmail.com>