Implement eval-able AnnotatedString 2-arg show (#54308)
The generic/fallback AbstractString 2-arg show omits the annotations,
meaning that eval(Meta.parse(repr(::AnnotatedString))) doesn't
round-trip.

The resolution to this is fairly simple, we just need to implement a
specialised show method. The implementation is fairly obvious, we're
just also able to get away with hiding the vector type annotation since
the constructor is fine without it.