Allow optimized OC in :new_opaque_closure (#54329)
This isn't officially supported, but some packages are currently using
the pattern of creating an optimized opaque closure, pulling out its
method and then using that in `:new_opaque_closure` to change the
environment. This stopped working properly when I changed optimized ocs
to not have a ->source field in the method. Let's keep this working for
now, but I'd like to provide a more first class mechanism for these
kinds of use cases in the near future.