Abbreviate `show(io, ::Module)`
For cases like

```
module A
    module B end
end
```

, `A` is already visible in `Main` once defined.
As a consequence, `B` can be accessed through `A.B`
without having to go through `Main`. This adjusts
the printing of `A.B`, so that this is reflected.

This also makes cases like

```
module A
    export B

    module B end
end
```

print like `B` and not `A.B` after `using A`, since
`B` is already accessible from `Main`.