Move compiled cache from DATAROOTDIR to LIBDIR
The FHS specifies that only architecture-independent files should be put
under /usr/share (DATAROOTDIR). Originally we only stored .ji files in
the "compiled" directory, but now we also put .so files.

Keep /usr/share and /usr/local/share in the default `DEPOT_PATH` for
backward compatibility and because they can still be used for
architecture-independent files such as environments or registries.

Also fix a few uses of `"data"` which should have been "`DATAROOTDIR`"
and wouldn't work if using a different `DATAROOTDIR` when building.