Revert "Move profiling manual section to a new tutorials toplevel heading (#52056)" (#54374)
This reverts commit 137783f1663ae0f7c1129c7d8031c874083b49fe.

I think there are many problems with the new Tutorial toplevel section:

- It should be very important since it adds a full new toplevel section
(to the existing five). However, it basically does not contain anything.
It was created by a somewhat arbitrary move of a manual-entry to the
tutorial section as well as a single page that more or less only
contains an external link.

- There is a non-obvious overlap between the Manual entries and the
Tutorial entries. The profiling entry was moved to tutorials but you
could take more or less any other entry in the manual entry and move it
as well.

- The comment in
https://github.com/JuliaLang/julia/pull/52056#pullrequestreview-1721618599
says
> We need a proper place for tutorials so tutorial-like PRs don't get
rejected and so that people can write actual tutorials content with the
correct audience in mind

While this might be true I don't think it is obvious at all that the
manual needs to be the place where people can "dump" general tutorials
about things. Right now, the Tutorial-section has more or less
"unbounded scope" and it isn't clear at all what should go in there.

- It has been there for 7 months without any real activity (except
taking a (again somewhat arbitrary) piece of the Pkg documentation into
it) so it seems it isn't really used for anything.

Considering all of the above, I think it is best to go back to the
status quo of 1.10 instead of "locking this in" for 1.11.

(cherry picked from commit 54dc748874c97fd38e3c795cfbcee46e4cca5cf0)